Relocated abstract classes
