+++ .github/tests/application/tests/TestCases/Common/AdminPagesTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\Common;
 
+use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\Abstracts\AbstractAdminPagesTest;
 use PHPUnit\Framework\Attributes;
 
 /**
